refactor: text errors to text messages (#476)
This patch implements a better way to deal with text messages by giving them a unique ID, a context, and a default message.

Closes #428

BREAKING CHANGES: Flows, request methods, form fields have had a key errors to show e.g. validation errors such as ("not an email address", "incorrect username/password", and so on. The `errors` key is now called `messages`. Each message now has a `type` which can be `error` or `info`, an `id` which can be used to translate messages, a `text` (which was previously errors[*].message). This affects all login, request, settings, and recovery flows and methods.